Renamed core.ts to base.ts
