+++ packages/core/agents/prompt-templates/react-prompt.ts
