+++ packages/core/agents/react.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
-import { Agent, AgentConfig } from '@/agent-core/agents/core';
-import { ModelResponse } from '@/agent-core/schema/core-schema';
-import { REACT_PROMPT } from '@/agent-core/agents/prompt-templates/react-prompt';
+import { Agent, AgentConfig } from '@/agent-core/agents/base';
+import { ModelResponse } from 'packages/core/schema/core-schema';
+import { REACT_PROMPT } from 'packages/core/agents/prompt-templates/react-prompt';
 
 export class ReactAgent extends Agent {
 	private maxIterations: number;
