+++ packages/core/memory/db/schema.ts
