+++ packages/core/readme.md
