+++ packages/core/schema/structured-output-schema.ts
