+++ packages/core/tools/readme.md
