+++ packages/examples/url-scrapper.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
-import { ReactAgent } from "@/agent-core/agents/react";
-import { WebScraperTool } from "@/agent-core/tools/web-scrapper";
+import { ReactAgent } from "packages/core/agents/react";
+import { WebScraperTool } from "packages/core/tools/web-scrapper";
 
 const OPENAI_API_KEY = 'openai_api_key';
